% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcctArbres.R
\name{AcctArbres}
\alias{AcctArbres}
\title{Evolution arbres}
\usage{
AcctArbres(df_Arbres)
}
\value{
La fonction analyse pour chaque arbrel'évolution
des principaux indicateurs, Gha, Vha, VcHa, VpHa, GainHa.
}
\description{
Evolution des principaux indicateurs par arbre.
}
\author{
Bruciamacchie Max
}
