% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcctPlacettes.R
\name{AcctPlacettes}
\alias{AcctPlacettes}
\title{Evolution Placettes}
\usage{
AcctPlacettes(df_Arbres)
}
\value{
La fonction analyse pour chaque placette l'évolution
des principaux indicateurs, Gha, Vha, VcHa, VpHa, GainHa.
}
\description{
Evolution des principaux indicateurs par placette
}
\author{
Bruciamacchie Max
}
