% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculBMPpcqm.R
\name{Calcul_BMP_pcqm}
\alias{Calcul_BMP_pcqm}
\title{Calcul du bois mort sur pied}
\usage{
Calcul_BMP_pcqm(df, seuilVisib = 25)
}
\value{
table BMPpcqm
}
\description{
Calcul du bois mort sur pied par échantillonnage linéaire.
}
\author{
Max Bruciamacchie
}
