% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcul_pcqm.R
\name{Calcul_pcqm}
\alias{Calcul_pcqm}
\title{Calcul PCQM}
\usage{
Calcul_pcqm(df, pop = "BMP", seuilVisib = 25)
}
\value{
table BMPpcqm
}
\description{
Calcul du bois mort sur pied par échantillonnage linéaire.
}
\author{
Demets Valentin
}
