% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcul_perches_pcqm.R
\name{Calcul_perches_pcqm}
\alias{Calcul_perches_pcqm}
\title{Calcul des perches}
\usage{
Calcul_perches_pcqm(df, seuilVisib = 25)
}
\value{
table Perchepcqm
}
\description{
Calcul de l'importance des perches avec l'échantillonnage PCQM.
}
\author{
Max Bruciamacchie
}
