% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_AcctD.R
\name{Calculs_AcctD}
\alias{Calculs_AcctD}
\title{Calcul des accroissements sur le diamètre}
\usage{
Calculs_AcctD(df_Arbres, Cycles)
}
\value{
La fonction complète le base arbres.
}
\description{
Pour la première mesure utilise la table AcctD, sinon calcule les accroissements réellement constatés.
}
\author{
Bruciamacchie Max
}
