% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_BMS_cercle.R
\name{Calculs_bms_cercle}
\alias{Calculs_bms_cercle}
\title{Calcul bois mort au sol}
\usage{
Calculs_bms_cercle(df = NULL, code_essreg = NULL)
}
\value{
Volume par hectare de bois mort au sol.
}
\description{
Calcul du bois mort au sol par cercle.
}
\author{
Demets Valentin
}
