% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_BMS_lineaire.R
\name{Calculs_BMS_lineaire}
\alias{Calculs_BMS_lineaire}
\title{Calcul bois mort au sol}
\usage{
Calculs_BMS_lineaire(df = NULL, code_essreg = NULL)
}
\value{
Volume par hectare de bois mort au sol.
}
\description{
Calcul du bois mort au sol par échantillonnage linéaire.
}
\author{
Bruciamacchie Max
}
