% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Gha.R
\name{Calculs_Gha}
\alias{Calculs_Gha}
\title{Calcul de Gha}
\usage{
Calculs_Gha(df_Arbres)
}
\value{
La fonction complète le base arbres.
}
\description{
fonction de calcul de la surface terrière/ha (feuille Arbres).
}
\author{
Bruciamacchie Max
}
