% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Ha.R
\name{Calculs_Ha}
\alias{Calculs_Ha}
\title{Calcul données /ha}
\usage{
Calculs_Ha(df_Arbres)
}
\value{
La fonction complète le base arbres.
}
\description{
Calcul des valeurs ramenées à l'hectare.
}
\author{
Bruciamacchie Max
}
