% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Miroirs.R
\name{Calculs_Miroirs}
\alias{Calculs_Miroirs}
\title{Traitement placettes miroirs}
\usage{
Calculs_Miroirs(df_Arbres, df_Echantillonnages)
}
\value{
La fonction complète le base arbres.
}
\description{
Ajout d'arbres sur les placettes miroirs.
}
\author{
Bruciamacchie Max
}
