% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Poids.R
\name{Calculs_Poids}
\alias{Calculs_Poids}
\title{Calcul de Nha (Poids)}
\usage{
Calculs_Poids(df_Arbres, df_Echantillonnages)
}
\value{
La fonction complète le base arbres.
}
\description{
fonction de calcul du poids (feuille Arbres).
}
\author{
Bruciamacchie Max
}
