% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Reges.R
\name{Calculs_Reges}
\alias{Calculs_Reges}
\title{Calculs régénération}
\usage{
Calculs_Reges(df)
}
\value{
La fonction prépare l'analyse de la régénération.
}
\description{
Importance de la régénération par essence et classe de hauteur.
}
\author{
Bruciamacchie Max
}
