% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Risques.R
\name{Calculs_Risques}
\alias{Calculs_Risques}
\title{Risque arbres.}
\usage{
Calculs_Risques(df_Arbres, Temps = 150, Rot = 8, tauxR = 0.0225)
}
\arguments{
\item{df_Arbres}{= Table Arbres}

\item{Temps}{= Temps de retour de l'aléa}

\item{Rot}{= Temps de rotation des coupes}

\item{tauxR}{= taux d'actualisation}

\item{df_Sauv}{= Valeurs de sauvegarde par essence et classe de diamètre}
}
\value{
La fonction construit la table Arbres.
}
\description{
Prise en compte des risques à l'échelle de l'arbre
}
\author{
Bruciamacchie Max
}
