% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Statuts.R
\name{Calculs_Statuts}
\alias{Calculs_Statuts}
\title{Calcul statuts}
\usage{
Calculs_Statuts(df_Arbres)
}
\value{
La fonction complète le base arbres.
}
\description{
Décomposition des tiges entre celles qui restent, celles coupées et les nouvelles.
}
\author{
Bruciamacchie Max
}
