% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Valeurs.R
\name{Calculs_Valeurs}
\alias{Calculs_Valeurs}
\title{Calcul valeurs}
\usage{
Calculs_Valeurs(df_Arbres, df_Quals, df_Prix, tauxR)
}
\value{
La fonction complète le base arbres.
}
\description{
Calcul des valeurs de consommation et potentielle.
}
\author{
Bruciamacchie Max
}
