% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculs_Volumes.R
\name{Calculs_Volumes}
\alias{Calculs_Volumes}
\title{Calcul des volumes.}
\usage{
Calculs_Volumes(df_Arbres, df_Tarifs)
}
\value{
La fonction complète le base arbres.
}
\description{
Calcul des volumes gestionnaires ou IFN à une ou deux entrées.
}
\author{
Bruciamacchie Max
}
