% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.r
\name{.create_color}
\alias{.create_color}
\title{Helper function to create color codes}
\usage{
.create_color(color, style = "normal")
}
\arguments{
\item{color}{Character ("cyan", "green", "white", 'black', 'purple', etc.)}

\item{style}{Character ("normal", "bold", "dim", "italic", "underline")}
}
\value{
ANSI escape code
}
\description{
Helper function to create color codes
}
\keyword{internal}
