% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.r
\name{.print_debug}
\alias{.print_debug}
\title{Helper function for nice debug/verbose printing}
\usage{
.print_debug(
  text,
  verbose = TRUE,
  debug = FALSE,
  type = "info",
  text_type = "INFO"
)
}
\description{
Helper function for nice debug/verbose printing
}
\keyword{internal}
