% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_AgreggPlac.R
\name{gf_AggregPlac}
\alias{gf_AggregPlac}
\title{Aggreg placettes.}
\usage{
gf_AggregPlac()
}
\value{
La fonction construit la table
}
\description{
Aggrégation des donnéese par placettes.
}
\author{
Bruciamacchie Max
}
