% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_AgregArbres.R
\name{gf_AgregArbres}
\alias{gf_AgregArbres}
\title{Aggrégation des arbres à l'échelle de la placette.}
\usage{
gf_AgregArbres(rep)
}
\value{
Les différentes tables sont enregistrées dans le dossier Tables
}
\description{
Sous-programme permettant d'aggréger toutes les variables à l'échelle de la placette.
Il fournit de nombreux tableaux croisés.
}
\author{
Bruciamacchie Max
}
