% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_AgregPlac.R
\name{gf_AgregPlac}
\alias{gf_AgregPlac}
\title{Aggreg placettes.}
\usage{
gf_AgregPlac()
}
\value{
La fonction construit la table
}
\description{
Aggrégation des donnéese par placettes.
}
\author{
Bruciamacchie Max
}
