% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_Calculs.R
\name{gf_Calculs}
\alias{gf_Calculs}
\title{Calcul des variables dendrométriques, économiques et écologiques.}
\usage{
gf_Calculs(tauxR)
}
\arguments{
\item{TauxR}{= taux d'actualisation (0.03 par défaut)}
}
\value{
La fonction construit la table Arbres.
}
\description{
Cette fonction commence par calculer le poids de chaque arbre,
puis pour chacun d'eux, calcule les variables dendrométriques, économiques et écologiques ramenées à l'hectare.
}
\author{
Bruciamacchie Max
}
