% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_Calculs_arbres.R
\name{gf_Calculs_arbres}
\alias{gf_Calculs_arbres}
\title{Calcul à l'échelle de l'arbre.}
\usage{
gf_Calculs_arbres(fich, tauxR)
}
\arguments{
\item{tauxR}{= taux d'actualisation}
}
\value{
La fonction construit la table Arbres.
}
\description{
Cette fonction commence par calculer le poids de chaque arbre,
puis pour chacun d'eux, calcule les variables dendrométriques, économiques et écologiques ramenées à l'hectare.
}
\author{
Bruciamacchie Max
}
