% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_Evol.R
\name{gf_Evol}
\alias{gf_Evol}
\title{Evolution des peuplemernts.}
\usage{
gf_Evol(df_Arbres)
}
\value{
La fonction construit la table Arbres.
}
\description{
Caractérisation de l'évolution des peuplements dès la première
remesure.
}
\author{
Bruciamacchie Max
}
