% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_Xls2Rdata.R
\name{gf_Xls2Rdata}
\alias{gf_Xls2Rdata}
\title{Import des données d'inventaire (modèle PermGF - utilisation locale)}
\usage{
gf_Xls2Rdata(repGF, fich)
}
\arguments{
\item{repGF}{= répertoire de travail}

\item{fich}{= classeur(s) excel(s) contenant les données d'inventaire}
}
\value{
Les données importées sont archivées au format .RData dans le fichier gfDonneesBrutes.RData
}
\description{
Permet d'importer les données d'inventaire contenues dans le classeur Excel de saisie.
}
\author{
Bruciamacchie Max, Demets Valentin
}
