% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphCompo.R
\name{GraphCompo}
\alias{GraphCompo}
\title{Composition en essence}
\usage{
GraphCompo(df1, df2)
}
\value{
Graphique ggplot.
}
\description{
fonction de calcul du poids (feuille Arbres).
}
\author{
Bruciamacchie Max
}
