% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNextractData.R
\name{IFNextractData}
\alias{IFNextractData}
\title{Import IFN data}
\usage{
IFNextractData(adress)
}
\value{
La fonction renvoie toutes les tables CSV de l'IFN.
}
\description{
Import des données brutes de l'IFN.
}
\author{
Bruciamacchie Max
}
